python_backend/requirements.txt
# Core dependencies
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.4
Pillow>=8.0.1
matplotlib>=3.3.3
scikit-image>=0.17.2

# Server dependencies
flask>=2.0.0
flask-cors>=3.0.10
gdown>=4.4.0

# Production dependencies
flask-compress>=1.10.0
gunicorn>=20.1.0

# Utilities
argparse>=1.4.0 